+++ serverless.yml
@@ -25,8 +25,8 @@ provider:
     SUMO_COLLECTOR_ID: ${file(./secrets/sumo.json):collectorId}
 
 functions:
-  lambdaHandler:
-    handler: dist/lambdaHandler.lambdaHandler
+  servePages:
+    handler: dist/servePages.servePages
     events:
       - http:
           method: ANY
@@ -34,6 +34,12 @@ functions:
       - http:
           method: ANY
           path: /{proxy+}
+  serveLogEndpoint:
+    handler: dist/serveLogEndpoint.serveLogEndpoint
+    events:
+      - http:
+          method: POST
+          path: /log
 
 package:
   excludeDevDependencies: false
